Add rewrap checking to build (#2416)
* Add rewrap checking to build

Unless I've misunderstood the discussion to date, we would prefer that
this be checked automatically so that we don't have a lot of "where do
these diffs come from? how are we supposed to prevent or fix them
again?" sorts of issues.

* Rewrap set.mm

* Do rewrapping on a temporary file

This might be just a cosmetic thing (the fact that I'm not
completely sure illustrates the point that mutation is hazardous) but
it seems better not to mutate set.mm and iset.mm themselves in the
rewrap check.

* Adjust scripts/rewrap order of operations

As suggested in CONTRIBUTING, run rewrap before save proof.

* Adjust CONTRIBUTING language about rewrapping

It is possible that most of the text introduced by "If you prefer
to run metamath yourself" can be removed or drastically pared
down, but at least for now maybe it is good to keep this for people who
are used to having it there? I don't think it is wrong per se.