Add a few theorems involving "if" to iset.mm (#2425)
* Add to missing theorems list in mmil.html

This is fvmptss , fvmptex , fvmptnf , fvmptn , fvmptss2 , fvopab4ndm ,
and fndmdifeq0 . The content of the entries is what had been in
mmil.html or a lightly edited version thereof.

* Add ifsbdc to iset.mm

The set.mm proof of ifsb works just fine if we add a condition that ph
is decidable and restate the theorem accordingly.

* Add fvifdc to iset.mm

This is fvif from set.mm with a decidability condition added. Other than
adding that condition, the theorem and proof are very similar to set.mm.

* Add ifeq1dadc to iset.mm

This is ifeq1da from set.mm with a decidability condition added - the
theorem and proof are similar to set.mm except for that.