Finish iset.mm proof of Bézout's identity (#2427)
* Add bezoutlemzz to iset.mm

This like bezoutlemex except that z can be any element of ZZ not just
NN0

* Add bezoutlemaz to iset.mm

This is like bezoutlemzz except that A can be an element of ZZ not just
NN0 .

* Add bezoutlembz to iset.mm

This is like bezoutlemaz except that B can be any element of ZZ , not
just NN0 .

* Add bezoutlembi to iset.mm

This is like bezoutlembz but with a biconditional instead of an
implication in the greatest common divisor part.

* Add bezoutlemmo to iset.mm

There is at most one nonnegative integer meeting the greatest
common divisor condition.

* Add bezoutlemeu to iset.mm

There is exactly one nonnegative integer meeting the greatest
common divisor condition.

* Add bezoutlemle to iset.mm

This will link supremum to the greatest common divisor condition we have
been using in the bezout lemmas.

* Add bezoutlemsup to iset.mm

This restates our greatest common divisor lemmas in terms of the
supremum notation.

* Add dfgcd3 to iset.mm

The theorem is stated as in set.mm. As in set.mm it relies on Bezout's
identity, or in this case relies on some of the lemmas for Bezout's
identity which we have proved so far.

* Add bezout to iset.mm

The theorem is stated as in set.mm. As described in the comment, the
proof is pretty different.

* Add Bezout's identity (iset.mm) to Metamath 100

It follows the same format as finds, which also is listed for both
set.mm and iset.mm.