Add least common multiple to iset.mm
This is the syntax lcm and df-lcm . We don't (currently) add _lcm and
df-lcmf because that is trickier.

Don't mention ex-lcm in df-lcm comment (we'll have it eventually, but
not yet).

Also add lcmval which is stated as in set.mm. The proof needs significant
changes to show that the infimum exists but other than that is similar
to the set.mm one.

Also mention ifcldadc in mmil.html because it won't be unusual to
need this rather than ifcldcd .