Add Euclid's Algorithm section to iset.mm (#2432)
* Add ifcldadc to iset.mm

This is ifclda from set.mm except that the condition of the if
expression has to be decidable.

* Add eucalgval2 to iset.mm

The theorem is stated as in set.mm. The proof is similar but needs some
changes around decidability of integer equality and related steps.

* Add eucalgval to iset.mm

Copied without change from set.mm.

* Add eucalgf to iset.mm

Copied without change from set.mm

* Add eucalginv to iset.mm

Copied without change from set.mm.

* Add eucalglt to iset.mm

The proof is stated as in set.mm. Although the overall structure of the
proof is based on the set.mm one, it needed some changes for real
versus rational, using closure instead of fvex , and rearranging to
avoid mt3d .

* Add eucialgcvga to iset.mm

This is eucalgcvga from set.mm with the usual iset.mm change to the
syntax of seq . The proof is taken from the set.mm proof with minor
changes needed for differences in algrf and algcvga .

* Add eucialg to iset.mm

This is eucalg from set.mm with the only differences being to change the
statement of the theorem and the proof for the different syntax of seq
in iset.mm and the (minor) differences in the algorithm theorems.

Add it to mmil.html and mm_100.html.