Add infssuzledc to iset.mm
This is like infssuzle in set.mm but significantly
changed to add conditions including around decidability.

Includes infssuzex which is a way of saying that the infimum exists.

Includes infuzre which is similar to set.mm's infssuzcl but
not as strong as the closure theorem we expect to prove eventually.