Add suplub2ti to iset.mm
This is suplub2 from set.mm with some changes. We have the same
tightness condition as supclti and other similar supremum theorems.
Unlike the other supremum theorems in this section, we also have a
condition that the order satisfies weak linearity (set.mm uses trichtomy
but weak linearity is enough). Other than those things the proof is
similar to the set.mm one.