Add least common multiple to iset.mm (#2436)
* Remove obsolete infimum comment in set.mm

This is in the section header at the start of least common multiple. It
is no longer true now that we have an infimum syntax and use it here.

* Add negiso to iset.mm

The theorem is stated as in set.mm. The proof is similar but needs some
adjustment around closure for -u .

* Add infrenegsupex to iset.mm

This is based on infrenegsup from set.mm and the main change needed, to
both the statement of the theorem and the proof, is how we express the
existence of the infimum. Other than that it is similar.

* Add supinfneg to iset.mm

This is part of what we'll need to convert between supremum and infimum
using -u negation.

* Add supminfex to iset.mm

This is supminf from set.mm with the usual changes for how we express
existence of a supremum or infimum.

* Add supelti to iset.mm

This will be useful in proofs which want to show that the supremum
is a member of a particular set. It is extracted from a portion of the
zsupcl proof.

* Shorten zsupcl proof by having it invoke supelti

* Add zssinfcl to iset.mm

This is analogous to zsssupcl and the proof is similar.

* Add infssuzledc to iset.mm

This is like infssuzle in set.mm but significantly
changed to add conditions including around decidability.

Includes infssuzex which is a way of saying that the infimum exists.

Includes infuzre which is similar to set.mm's infssuzcl but
not as strong as the closure theorem we expect to prove eventually.

* Update mmil.html supremum entries

This is for suprzcl , suprleub , and suprzub . We have partial
equivalents or thoughts about what might be possible.

* Add zdvdsdc to iset.mm

We already had dvdsdc which said that divisibility was decidable if the
arguments are in NN and ZZ respectively. This extends it to ZZ and ZZ.
The proof is straightforward from integer trichotomy and divisibility
theorems we already have.

* Add least common multiple to iset.mm

This is the syntax lcm and df-lcm . We don't (currently) add _lcm and
df-lcmf because that is trickier.

Don't mention ex-lcm in df-lcm comment (we'll have it eventually, but
not yet).

Also add lcmval which is stated as in set.mm. The proof needs significant
changes to show that the infimum exists but other than that is similar
to the set.mm one.

Also mention ifcldadc in mmil.html because it won't be unusual to
need this rather than ifcldcd .