Add lcmcllem to iset.mm
The theorem is stated as in set.mm. The proof is different from the
set.mm one in many details - for example the decidability hypotheses
of infssuzcldc and stating that not equal is equivalent to apart for
integers - but is similar in the sense of showing that
( abs ` ( M x. N ) ) is an upper bound for the least common multiple
and using that to show that the least common multiple exists.