Add 3lcm2e6woprm to iset.mm
The proof is stated as in set.mm. The proof is almost the same as the
set.mm one but changes not equal to apart in one set of steps.

Modify the comment to not refer to 3lcm2e6 which we won't be able to
prove for a while (at least not with the proof which refers to
primality).

Do not copy over the discouraged tags from set.mm at least until we
have 3lcm2e6 .