Add ncoprmgcdne1b to iset.mm
The theorem is stated as in set.mm. The proof is rewritten, because it
isn't clear whether any local repair is possible for the use of excluded
middle in the set.mm proof. Fortunately, the rewritten proof isn't hard.
It just proves each direction of the biconditional separately and proceeds
fairly straightforwardly from gcd and || theorems we already have.