Add 1nprm to iset.mm
The theorem is stated as in set.mm. The proof is similar to the set.mm
except at the end where the set.mm one relies on strict dominance and
php4 and this proof uses php5 instead (in either case it is trying to
show that a class cannot be equinumerous to both 1o and 2o ).