Add pm54.43 to iset.mm
The theorem is stated as in set.mm.  The proof is almost the same except
that 1sdom2 is changed to 1nen2 .

The set.mm comment is mostly deleted - even between Principia and set.mm there
are a fair number of differences in terms of things like cardinality,
and cardinality changes in iset.mm yet more, so it seems best to leave
this discussion in set.mm and not try to adapt it here.