Add prime numbers to iset.mm through dvdsprm (#2451)
* Add Prime to iset.mm

This is the Prime syntax and the df-prm definition.

* Add isprm to iset.mm

Copied without change from set.mm.

* Add prmnn and prmz to iset.mm

Copied without change from set.mm.

* Add prmssnn and prmex to iset.mm

Copied without change from set.mm.

* Add 1nprm to iset.mm

The theorem is stated as in set.mm. The proof is similar to the set.mm
except at the end where the set.mm one relies on strict dominance and
php4 and this proof uses php5 instead (in either case it is trying to
show that a class cannot be equinumerous to both 1o and 2o ).

* Add 1idssfct to iset.mm

Copied without change from set.mm.

* Add prfi to mmil.html

I'm reluctant to say for sure it cannot be proven but it sort of looks
like it cannot.

* Add onirri to iset.mm

Copied basically without change from set.mm (although it is later in the
file and with a renamed hypothesis).

* Add 1nen2 to iset.mm

This is like 1sdom2 but we haven't (yet at least) defined a strict
dominance syntax in iset.mm. Use it in the 1nprm proof.

* Mark elirr as discouraged.

We'd like people to use ordirr when possible (which is probably most of
the time).

* Add pm54.43 to iset.mm

The theorem is stated as in set.mm.  The proof is almost the same except
that 1sdom2 is changed to 1nen2 .

The set.mm comment is mostly deleted - even between Principia and set.mm there
are a fair number of differences in terms of things like cardinality,
and cardinality changes in iset.mm yet more, so it seems best to leave
this discussion in set.mm and not try to adapt it here.

* Add pr2nelem to iset.mm

Copied without change from set.mm.

* Add pr2ne to iset.mm

The theorem is stated as in set.mm. The proof is bascially the set.mm
one but needs (1) to change 1sdom2 to 1nen2 , and (2) to change prex to
prexg and rearrange accordingly.

* Add mt4d to mmil.html

This is yet another ax-3 like theorem.

* Add en2eqpr to iset.mm

The theorem is stated as in set.mm. The proof is totally different
because I couldn't find a proof for fisseneq but because en2eqpr only
deals with two element sets it is feasible to just list all the cases
for how each set can be mapped with 2o .

Mention en2eqpr in mmil.html entry for fisseneq .

* Add isprm2lem to iset.mm

The theorem is stated as in set.mm. The proof is from en2eqpr - the
current set.mm proof uses excluded middle.

* Shorten isprm2lem proof in set.mm

Copied from the one I just wrote for iset.mm. This takes the proof from
1030 characters to 570 characters.

* Add isprm2 to iset.mm

Copied without change from set.mm.

* Add isprm3 to iset.mm

The theorem is stated as in set.mm. The proof is basically the set.mm
one but with a few changes for not equal versus apart and decidability
of integer equality.

* Add isprm4 to iset.mm

The theorem is stated as in set.mm. The proof is basically the set.mm
one but needs some adjustments to use decidability of integer equality
and rearrange the proof accordingly.

* Add prmind2 and prmind to mmil.html

It would seem like this can be proved but we haven't proved them yet.

* Add dvdsprime to iset.mm

Copied without change from set.mm.

* Add nprm to iset.mm

Copied without change from set.mm.

* Add nprmi to iset.mm

Copied without change from set.mm.

* Add dvdsnprmd to iset.mm

Copied without change from set.mm.

* Mention dfordc in mmil.html

When dealing with integers, for example, decidability can be
highly relevant.

* Add prm2orodd to iset.mm

The theorem is stated as in set.mm. The proof is the set.mm one except
that the equivalent to orrd needs decidability of integer equality.

* Add 2prm to iset.mm

Copied without change from set.mm.

* Add 3prm to iset.mm

Copied without change from set.mm.

* Add 4nprm to iset.mm

Copied without change from set.mm.

* Add prmuz2 to iset.mm

Copied without change from set.mm.

* Add prmgt1 to iset.mm

Copied without change from set.mm.

* Add prmm2nn0 to iset.mm

Copied without change from set.mm.

* Add oddprmgt2 to iset.mm

The theorem is stated as in set.mm. The proof is the set.mm one except
that ltlen needs to be changed to zltlen .

* Add oddprmge3 to iset.mm

Copied without change from set.mm.

* Add sqnprm to iset.mm

Copied without change from set.mm.

* Add dvdsprm to iset.mm

Copied without change from set.mm.