Preparatory theorems 1/2: the 11 el*v* theorems
I submit my Preparatory theorems in 2 parts: this is the controversial one. I think these 11 theorems are ok in my Mathbox, and they would decrease the size of set.mm with hundreds of theorems with  ` $e |- A e. _V ` , but I’m quite sure I can’t see the playing field completely. I’ll just write here what I do:
When I met with a ` $e |- A e. _V ` , I found the general version and I included only this general version in my Mathbox. I don’t use the ` g ` suffix at all because the general version is the default form in my Mathbox (of course I use ` d ` and ` f ` versions). In proofs, I use the default general version this way: ` elv.1 $e |- ( x e. _V -> ph ) $. `  ` elv $p |- ph $= ` . Of course, exceptions exist, and
` A ` is not always  ` x ` in proofs (for example, I used inference form ~ elqs with ~ 0ex and not the general form ~ elqsg in my ~ n0elqs in order to minimize it),
and if you need the same ` $e |- x e. _V ` in a proof several times, the proof can be longer,
but I find this approach by and large simpler and more economical.