Add expcan to iset.mm
The theorem is stated as in set.mm. The proof in general terms follows
the same logic as the set.mm proof but the details are quite different,
starting with the reliance on letri3 and lenlt rather than real number
trichotomy, and continuing with the choice to just break out one
direction of the order into the lemma expcanlem rather than proving
and using eqord1 .

Adjust mmil.html comment for eqord1 accordingly.