Add fundmfibi to iset.mm
The theorem is stated as in set.mm. The proof is similar but adjusted to
use fundmfi instead of dmfi .

The comment is from AV and removes the word (set). It was intended to convey
that F is also a set but seems unnecessary as discussed on github with AV and BJ.
AV will be making that change to set.mm.