Add maxltsup to iset.mm
The statement of the theorem is maxlt from set.mm with maximum expressed in
terms of supremum. The proof of the forward direction is similar to the
set.mm proof. The proof of the reverse direction does not follow the
set.mm one but instead is in terms of absmax and absltd .