Add set size function to iset.mm (#2525)
* Add sumeq2 to iset.mm

The theorem is stated as in set.mm. The proof is similar to the sumeq2d
one (but if there was a way to prove it in terms of subeq2d I didn't
find it, at least as the distinct variable constraints are arranged at
the moment).

* Add fisupg to mmil.html

* Add infm and infn0 to iset.mm

infn0 is stated as in set.mm. Although the set.mm proof doesn't work,
there is an easy proof from brdomi .

infm is the same thing but with not empty changed to inhabited.

* Add NN0* to iset.mm

This is df-xnn0 and the NN0* syntax. It is copied without change from
set.mm except for the removal of references in comments to theorems
which iset.mm does not yet have.

* Move extended real theorems earlier in iset.mm

This is pnfxr , pnfex , pnfnemnf , mnfnepnf , and mnfxr . The new
position is between renemnfd and rexri which is the same position they
have in set.mm.

* Add rest of the NN0* section to iset.mm

Copied without change from set.mm.

This is elxnn0 , nn0ssxnn0 , nn0xnn0 , xnn0xr , 0xnn0 , pnf0xnn0 ,
nn0nepnf , nn0xnn0d , nn0nepnfd , xnn0nemnf , xnn0xrnemnf , and
xnn0nnn0pnf .

* Add xpfi to mmil.html

Looks like this one is possible.

* Add znnen and qnnen to mmil.html

Looks like these would be provable.

* Add set size to iset.mm

We use "size" as the ASCII version of the function and a sharp sign as
the typeset version.

The definition should preserve the usual properties for finite sets and
infinite sets (where we define A to be infinite if _om ~<_ A ).

* Add infinf to mmil.html

This gets at the heart of how to define "A is infinite" so add a few
sentences on that.

* Clarify domtriom comment in set.mm

Although I suppose it is kinda clear from context that CC is countable
choice rather than (say) the complex numbers, it seems better to be
explicit.

* Add fict to iset.mm

The theorem is stated as in set.mm. The proof is vaguely similar to the
set.mm one at least in the sense of relying on ssdomg but in the set.mm
case that's a few lemmas down and here is it more direct.

* Add sizeinfuni to iset.mm

This will be a part of showing that df-size works as expected.

* Add hashkf , hashgval , hashginv to mmil.html

* Add sizeinf to iset.mm

The statement of the theorem is slightly changed from set.mm. The proof
is different because of the differences between set.mm and iset.mm with
respect to set size.

* Add hashbnd to mmil.mm

* Add hashf and friends to mmil.html

This is hashfxnn0 , hashf , hashxnn0 , hashresfn , dmhashres ,
and hashnn0pnf .

* Add infnfi to iset.mm

This is one direction of infinf from set.mm. The proof is from
php5dom and other theorems we have.

* Add hashnnn0genn0 to mmil.html

* Add hashnemnf to mmil.html.

* Add sizeennnuni to iset.mm

This is the counterpart to sizeinfuni for finite sets (except that it
also specifies the specific size).

* Add sizecl to iset.mm

This is a finite set counterpart to sizeinf . It corresponds to hashcl
in set.mm.

Includes sizeennn which takes sizeennnuni and takes it one step further,
to express the same idea in terms of the size function (and the usual
mapping between _om and NN0 ).

* Move fz01or earlier in iset.mm

There's no reason not to put it in the main section with df-fz and
various theorems involving finite integer ranges.

* Add sizefiv01gt1 to iset.mm

This is hashv01gt1 from set.mm but for the finite set case.

* Add sizefz1 iset.mm

This is hashfz1 from set.mm with the set size function renamed. The
proof is from sizeennn unlike the set.mm proof.

* Add sizeen to iset.mm

This is hashen from set.mm with a rename to the set size function. The
proof is new rather than being taken from the set.mm one.

* Add hasheni to mmil.html

* Add sizeeqf1o to iset.mm

This is hasheqf1o from set.mm. The only change in the statement of the
theorem or the proof is the name of the set size function.

* Add f1ofi to iset.mm

This is a simple consequence of theorems we already have and might not
be worth its own theorem except that it seems worth making the parallel
to fofi in set.mm. Add fofi entry to mmil.html.

* Add fiinfnf1o to iset.mm

The theorem is stated as in set.mm. The proof is similar to the set.mm
but fofi needs to be changed to f1ofi . In the comment, also discuss
what "infite set" means with reference to infnfi .

* Add focdmex to iset.mm

Copied without change from set.mm.

* Add sizeeqf1oi to iset.mm

This is basically hasheqf1oi from set.mm but for finite sets (and with a
bit of rearrangement which shouldn't change anything).

* Add sizef1rn to iset.mm

This is hashf1rn from set.mm but restricted to finite sets. The proof is
similar to the set.mm one with some small adjustments.

* Add sizeeqf1od to iset.mm

This is similar to hasheqf1od from set.mm but for finite sets. The proof
is trivial from sizeeqf1oi .

* Add fz1eqb to iset.mm

The theorem is stated as in set.mm. The only difference in the proof is
the name of the set size function but other than that the proof is the
same as the set.mm one.

* Add hashcard to mmil.html

* Add hashcl to mmil.html

This corresponds to sizecl in iset.mm except for the name of the set
size function.

* Add hashxrcl , hashclb , hashvnfin to mmil.html

Seems like we probably won't get these ones.

* Add filtinf to iset.mm

This is a bit similar to nfile in set.mm but with some differences.
Note one of the differences in a new entry for mmil.html.

* Add hashvnfin and hashnfinnn0 to mmil.html