The union of two finite sets is finite if their intersection is (iset.mm) (#2537)
* Add unfiin to iset.mm

This is similar to unfidisj but the sets don't need to be disjoint, they
just need to have a finite intersection. Proposed by Mario Carneiro, who
also suggested using undiffi and diffifi in the proof, but beyond those
hints I wrote the proof.

* Expand infi entry in mmil.html

This was suggested by Mario Carneiro in a github comment.