Change X e. On to Ord X in tfr1on
This is in tfr1on itself and most of its lemmas.

I had always thought of the condition on X as being Ord X so I'm not
sure why I wrote down X e. On when I started writing these proofs.
The changes needed to switch to Ord X are not large - ontr1 changes
to ordtr1 and the like, but nothing had been expecting X to be a set.