Add tfrcl to iset.mm
Closure theorem for recursion.

This is similar to tfr1on but has a hypothesis saying that the recursion
rule stays within a given set S , and the conclusion says that the
value of the function produced by recursion is in that set.

Includes lemmas tfrcllemsucfn , tfrcllemsucaccv , tfrcllembacc ,
tfrcllembxssdm , tfrcllemssrecs , tfrcllembfn , tfrcllembex ,
tfrcllemubacc , tfrcllemex , tfrcllemaccex , and tfrcllemres
which are similar to the similarly named lemmas in the tfr1on proof.

Includes tfrcldm which states that the result of recursion (given this
kind of closure hypothesis) is defined up to a point based on where the
closure hypothesis is valid.