Revise freccl in iset.mm
Rename the current freccl to frecclOLD . There is only one usage but we
don't need to move that over yet.

Add freccl which is the same as frecclOLD but which has one fewer
hypotheses. Prove it in terms of tfrcl (and put most of the proof into a
lemma freccllem so that we can give a name to a common expression).