Add frecuzrdgtclt to iset.mm
This is a version of frecuzrdgtcl but where the set which we have
closure theorems for ( S ) need not be the set used in defining the
function we pass to frec but merely a subset of it.

Includes frecuzrdgg which connects the two kinds of natural number used
in the recursive definition generator on upper integers.

Includes frecuzrdgdom and frecuzrdgfun which break out portions of
the frecuzrdgtclt proof (and I suppose could be useful on their own too).