Remove hypothesis from frecuzrdgdom and friends
This is frecuzrdgdom , frecuzrdgfun , frecuzrdgsuct . In all three cases
there had been a hypothesis for the definition of G (the mapping between
_om and ( ZZ>= ` C ) ) which had not been used in the final theorem,
only in the proof. Move that hypothesis, and the bulk of the proof,
to lemmas frecuzrdgdomlem , frecuzrdgfunlem , and frecuzrdgsuctlem ,
respectively, and redefine frecuzrdgdom , frecuzrdgfun , and frecuzrdgsuct
to not require that hypothesis.