Add some ordinal theroems to mmil.html
These are theorems, mostly about ordinal addition and multiplication,
which are not present in iset.mm but had not been listed in the missing
theorems list in mmil.html.  In most cases we can only say that the
set.mm proof doesn't work; for oawordri we can also say that the theorem
would imply excluded middle.

Remove nnawordex from the missing theorems list as it is now present in
iset.mm.