Change typesetting of # in set.mm
This changes the # (set size) function in set.mm to typeset as a sharp
sign. This is for consistency with iset.mm. Although it could be
justified in more abstract terms by saying that the octothorpe most
conventionally represents an apartness relation, that's an argument
which sort of only makes sense when comparing with iset.mm, because
set.mm has no need for an apartness relation other than =/= (not equal).