Add phicl2 to iset.mm
The theorem is stated as in set.mm. The proof is similar to the set.mm
but needs intuitionizing on various steps around showing that various
sets exist or are finite.

Also add phivalfi which breaks out part of the phival proof into a
separate theorem we can use elsewhere.