Universal property of the disjoint union
As discussed in PR #2636, the universal property of the disjoint union is provided and proven: ~updjud. In contrast to the prposal in PR #2636, this theorem holds for the disjoint union of sets only, not for classes in general (see hypotheses updjud.a and updjud.b). Furthermore, the very general definitions of ` inl ` and ` inr ` had ro be restricted to the corresponding (base) sets of the disjoint union.

To prove this theorem, several auxiliary theorems have been provided.