Add exmidfodomr to iset.mm
Includes exmidfodomrlemim which is the direction from fodomr in set.mm.
The proof is basically the set.mm proof of fodomr, rearranged a bit to
invoke EXMID where needed.

Includes lemma exmidfodomrlemr which is the direction showing that
fodomr implies excluded middle.

Includes lemma exmidfodomrlemeldju which is a variation of djur ,
somewhat specific to this proof.

Modify fodomr entry in mmil.html accordingly.