Add notnotsnex to iset.mm
The statement of the theorem and proof are taken from a github comment
by Mario Carneiro; I just put it into metamath format.

Add a discussion of the provability of { A } e. _V in iset.mm. This is
also taken from a github comment by Mario Carneiro and lightly edited.
Since we don't have a way to formalize this argument (at least
that we have found so far), I'll call it a "conjecture".