Fix nfof and nfofr
They both had a distinct variable constraint which defeated the point of
having such theorems. The change is to remove that constraint and adjust
the proofs accordingly.

This is for both set.mm and iset.mm and the new proofs are the same
except for one occurence of nfral vs nfralxy .