proofs containing 2pos shortened (#2683)
* proofs containing 2pos shortened

Inspired by the discussion about finding lemmata which minimize set.mm (see Google group), a lot of prrofs can be shortened by using theorems based on the fact that 2 e. RR+.  This almost always removes 2pos from the proof. The following proofs are affected (all theorems up to and including PART 6 are checked):
~zgt1rpn0n1, ~2tnp1ge0ge0, ~fldiv4lem1div2uz2, ~2swrd2eqwrdeq, ~sqrlem7, ~amgm2, ~iseralt, ~climcndslem2, ~climcnds, ~geoihalfsum, ~efcllem, ~oddge22np1, ~nn0ehalf, ~nno, ~nn0oddm1d2, ~flodddiv4t2lthalf, ~oddprm, ~iserodd, ~prmgaplem7,

Furthermore, two theorems ~prodge0d and ~prodge02d were added, which are deduction versions of existing theorems, and also using the  membership in RR+. with this, the proofs of the following theorems could be shortened:
~evennn02n, ~nvge0, ~ oexpneg, ~oexpnegALTV, ~pjthlem1, ~pjhthlem1

* prodge0, prodge0i, prodge02 obsolete