Superfluous hypotheses removed (4)+(5); deduction versions for inferences; generalization of map1 (#2694)
* Superfluous hypotheses removed (3)

~lssneln0 revised by removing a superfluous hypothesis according to the list of @savask, see issue #2648.

Theorems depending on this theorem (mostly in Norm's Mathbox) were revised.  There is one theorem (hmapcl) which is commented out and using ~lssneln0. I left ~lssneln0OLD within its proof, but I think it is very difficult to keep such theorems/proofs maintained in the long run.

* Deduction versions for inferences; generalization of map1

* ~mapsnd and ~mapsnend moved from GS's mathbox to main set.mm
* proofs of ~mapsnd and ~mapsnend shortended using ~mapsnd and ~mapsnend
* generalizations ~snmapen and ~snmapen1 of ~map1 added, proof of ~map1 shortened
* proofs shortened using ~mapsnend instead of ~mapsnen: ~map2xp, ~mapdom3, ~ackbij1lem5, ~pwxpndom2, ~hashmap

* Comments, Rewrap

Comments revised: the Contributor of a new theorem in deduction form should  be the contributor of the already existing inference form (especially if the old proof is only slightly modified). The author of the new theorem should be mentioned in the "revised by" tag.

* Superfluous hypotheses removed (5)

~wfrlem4 (contributed by SF) revised by removing a superfluous hypothesis according to the list of @savask, see issue #2648.

* Revision lssneln0

* new theorem ~ lssvneln0 (part of ~lssneln0)
* proof of ~lssneln0 shortened using lssvneln0
* proofs of ~lvecindp, ~lshpne0 shortened using ~lssvneln0