First cut at documentation page for $j annotations
This is text by Jim Kingdon and Mario Carneiro from a github issue.
The formatting is pretty rough and the content might be too.
But it is at least a start.

Include paragraph which tries to get at how "official" this list is.

Include list of tools which currently parse $j comments.