Tweaks to wording
First, speak of $j commands and drop the word "annotations" (which might
be an OK word but which isn't really needed when we already have other
words for them).

Secondly, try another pass at the paragraph about what this document is.
Hopefully we can get something which is close enough.