Add nnsf to mathbox
This is defining a function S (which will be used in various parts of
the proof that NN+oo is omniscient) and showing that it has the expected
domain and range.

Includes nnpredcl which is similar to theorems like onsucuni2 but we don't
have it in quite this form.

Includes 0lt2o and 1lt2o .  These are convenience theorems which are
frequently used when we are treating 2o as a boolean type.

Includes nnsucpred which is a theorem about successors and predecessors
of natural numbers.