Add isbth to iset.mm
Add sbthlem1 to iset.mm
Copied without change from set.mm.

Add sbthlem2 to iset.mm
Copied without change from set.mm.

Add sbthlemi3 to iset.mm
This is sbthlem3 from set.mm with the EXMID condition added. The proof
is similar to the set.mm one with that condition added as needed.

Add sbthlemi4 to iset.mm
This is sbthlem4 from set.mm with an EXMID condition added. The proof is
the same as the set.mm one, merely changed to pass EXMID to sbthlemi3 .

Add sbthlemi5 to iset.mm
This is sbthlem5 from set.mm. It is changed to add an EXMID condition,
which is needed to invoke sbthlemi3 and undifdcss (which replace
sbthlem3 and undif , respectively).

Add sbthlemi6 to iset.mm
This is sbthlem6 from set.mm with an EXMID condition added. The proof is
the same as the set.mm one modified to pass the EXMID condition on to
sbthlemi4 and undifdcss .

Add sbthlem7 to iset.mm
Copied without change from set.mm.

Add sbthlemi8 to iset.mm
This is sbthlem8 from set.mm with an EXMID condition added. The proof is
the set.mm one with changes to pass the EXMID condition along to sbthlemi4 .

Add sbthlemi9 to iset.mm
This is sbthlem9 from set.mm with an EXMID condition added. The proof is
basically the set.mm one but rearranged to pass EXMID to the various lemmas.

Add sbthlemi10 to iset.mm
This is sbthlem10 from set.mm with an EXMID condition added. The proof
is only slightly changed from the set.mm one - to pass the EXMID
condition along to sbthlemi9 .

Add isbth to iset.mm
This is sbth from set.mm with the addition of an EXMID condition. The
proof is similar to the set.mm one with adjustments to add the EXMID
condition as needed.