Alternate definition of iota
As proposed by BJ via email on 13-Aug-2022, an alternate definition of iota is introduced (in AV's mathbox) which is the universe V if it is not defined (in contrast to the current definition, which is the empty set in this case). By this definition, the alternate definition of a function value can be adapted accordingly (in a next PR).

Changes in detail (main set.mm):
* ~elv moved from PM's mathbox (TODO: check if additional proofs can be shortened)
* ~dfiota2lem extracted from proof of ~dfiota2
* new symbol "iota'"

Changes in detail, AV's mathbox:
* new section "General auxiliary theorems (1)" with auxiliary theorems created (some theorems of section "General auxiliary theorems (2)" moved here)
* new auxiliary theorems ~al0ssb, ~eusnsn, ~absnsb, ~euabsn22
* new section "Definite description binder (inverted iota) - extension" with two examples for the current iota.
* new section "Alternative for Russell's definition of a description binder" with the definition of the alternate iota and basic theorems for it.