Add iseqcoll to iset.mm
This is basically seqcoll from set.mm but with the usual changes for
differences in seq including hypotheses that F and H be defined beyond
where they "need" to be. The proof is taken from the set.mm proof with
the changes being to adjust the invocation of various seq related
theorems.