Remove distinct variable constraint from isumrb
There had been a constraint between B and k not just in isumrb but also
isummo and a bunch of lemmas. This is unfortunate since the whole point
of B is to have a k in it. However, this was basically an oversight in
a few of the steps in isumrblem (not the rest of the proofs in this
area) and was easily enough corrected by using an expression
like [_ n / k ]_ B or similar in the appropriate places.